requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
faiss-cpu
python-multipart
